move store.ts to lib dir
