+++ apps/tui/src/lib/store.ts
